aaaimx-project/requirements.txt
asgiref==3.3.1
Django==3.1.6
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
importlib-metadata==3.4.0
Markdown==3.3.3
Pillow==8.1.0
pkg-resources==0.0.0
PyJWT==2.0.1
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
zipp==3.4.0
